back-end of UI overhaul
 * windows will no longer disappear and reappear
 * reduced number of files and overall size of the app
 * renamed main file to main.py
 * updated tut.sh for Drauger OS 7.5.1
