+++ usr/share/applications/Drauger_Welcome.desktop
@@ -1,7 +1,7 @@
 [Desktop Entry]
 Name=Drauger Welcome
 Version=1.0
-Exec=/usr/share/drauger-welcome/welcome.py
+Exec=/usr/share/drauger-welcome/main.py
 Comment=Welcome Screen for Drauger OS
 Icon=/usr/share/icons/Drauger/scalable/menus/drauger_os-logo.svg
 Type=Application
